\name{clean.events}
\alias{clean.events}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
clean.events
}
\description{
A helper function for \code{\link{longCatEDA}} which cleans out missing data from \code{events} and \code{event.times}.
}
\usage{
clean.events(events, event.times)
}
\arguments{
  \item{events}{
see \code{\link{longCatEDA}}
}
  \item{event.times}{
see \code{\link{longCatEDA}}}
}
\author{
Stephen Tueller
}
