% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example2cat}
\alias{example2cat}
\title{Example data for \code{\link{longCatPlot}}}
\format{
Data frame with 20 subjects and 6 time points
}
\source{
Tueller, S. J., Van Dorn, R. A., & Bobashev, G. V. (2016). Visualization of categorical longitudinal and times series data (Report No. MR-0033-1602). Research Triangle Park, NC: RTI Press.
}
\usage{
example2cat
}
\description{
Simulated data for illustrating \code{\link{longCatPlot}}.
}
\keyword{datasets}
