% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longContPlot.R
\docType{data}
\name{example2cont}
\alias{example2cont}
\title{Simulated Data for Illustrating \code{\link{longContPlot}}}
\format{
Data frame with 20 subjects and 6 time points
}
\usage{
example2cont
}
\description{
Simulated data for illustrating \code{\link{longContPlot}}.
}
\keyword{datasets}
