\name{example3}
\docType{data}
\alias{example3}
\title{Simulated Data for Illustrating longCat and longCatPlot}
\description{
Simulated data for illustrating \code{\link{longCat}} and \code{\link{longCatPlot}}.
}
\usage{example3}
\format{Data frame with 100 subjects and 6 time points}
\references{
Tueller, S. J., {van Dorn}, R., and Bobashev, G., (2013). Visualization of Categorical Longitudinal and Times Series Data.
}
\keyword{datasets}
