\name{levelCheck}
\alias{levelCheck}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function to Check Factor Levels
}
\description{
Internal function of checking the levels of \code{y} called by \code{\link{longCat}}.
}
\usage{
levelCheck(y)
}
\arguments{
  \item{y}{a data matrix, see documentation for \code{y} in \code{\link{longCat}}.}
}
\references{
Tueller, S. J., Van Dorn, R. A., and Bobashev, G. V.  (2013). Visualization of Categorical Longitudinal and Times Series Data. Manuscript Under Review.}
\author{Stephen Tueller}

\seealso{
\code{\link{longCat}}.
}

