\name{lunique}
\alias{lunique}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Find the Unique Number of Factors
}
\description{
Function to find the unique number of factors in a matrix or data frame of categorical data. Used internally by \code{\link{longCat}}.
}
\usage{
lunique(y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{y}{see \code{y} for \code{\link{longCat}}.
}
}

\references{
Tueller, S. J., Van Dorn, R. A., and Bobashev, G. V.  (2013). Visualization of Categorical Longitudinal and Times Series Data. Manuscript Under Review.}
\author{Stephen Tueller}


