% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorter.R
\name{makePatterns}
\alias{makePatterns}
\title{Concatenate Multivariate Data into Numeric or Character Patterns}
\usage{
makePatterns(dat, times = NULL, num = TRUE, mindur = NULL, igrpt = FALSE)
}
\arguments{
\item{dat}{A matrix or data frame such as \code{lc$y} from an \code{\link{longCat}} object created by \code{\link{longCat}}.}

\item{times}{See \code{times} in \code{\link{longCat}}.}

\item{num}{Logical indicator, specifying whether a numeric version of the concatenated rows should be returned. When \code{TRUE}, the return is rescaled by moving a decimal point between the first and second digits. This ensures that, under different numbers of observations or missing data, ordering is not unduly impacted by patterns of missing data. Default is \code{TRUE}.}

\item{mindur}{Minimum duration. If \code{times} is a matrix or data frame of individually varying times of observation of the same dimension as \code{dat}, selecting \code{mindur > 0} results in all cells in \code{y} corresponding to cells in \code{times - times[,1] < mindur} being changed to \code{NA}. This minimizes the effect of short durations on the sorting algorithm in \code{\link{sorter}}. Default is \code{NULL}.}

\item{igrpt}{Option to ignore repeated values when sorting, allowing the sorting algorithm in \code{\link{sorter}} to smooth over regions of no change for each row in \code{lc$y}. Default is \code{FALSE}.}
}
\value{
A vector of patterns of length \code{nrow(dat)}.
}
\description{
This function concatenates the columns of a matrix or data frame for each row into a single character variable, which can optionally be reconverted to numeric. It is typically called internally by \code{\link{sorter}}. For example, a row of a matrix containing \code{c(1, 2, 3, 5)} will be concatenated into \verb{"1235"}.
}
\examples{
# create an arbitrary matrix and demonstrate
temp <- matrix( sample(1:9, 40, replace=TRUE), 10, 4)
print(temp)
makePatterns(temp, num=FALSE)

# examine the unique patterns of data
bindat <- matrix( sample(0:1, 500, replace=TRUE), 100, 5)
uniquePatterns <- makePatterns(bindat, num=FALSE)
as.matrix(table(uniquePatterns))

}
\references{
Tueller, S. J., Van Dorn, R. A., & Bobashev, G. V. (2016). Visualization of categorical longitudinal and times series data (Report No. MR-0033-1602). Research Triangle Park, NC: RTI Press. \url{http://www.rti.org/publication/visualization-categorical-longitudinal-and-times-series-data}
}
\seealso{
\code{\link{sorter}} for further processing of patterns.
}
\author{
Stephen J. Tueller
}
