% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorter.R
\name{norpt}
\alias{norpt}
\title{A function to take out repeated observations in a vector of data for sorting purposes.}
\usage{
norpt(alist = c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4, 5))
}
\arguments{
\item{alist}{A numeric vector from which repeated values are to be removed.}
}
\value{
A vector excluding repeated values with trailing \code{NA}'s to fill the vector to its original length.
}
\description{
This function removes repeated observations from a vector, filling the "gaps" with \code{NA} values to maintain the original vector length. It's useful for preparing data for sorting where unique values are necessary, and the overall structure of the dataset (i.e., its length) should be preserved.
}
\examples{
alist <- c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4, 5)
norpt(alist)

}
\references{
Tueller, S. J., Van Dorn, R. A., & Bobashev, G. V. (2016). Visualization of categorical longitudinal and times series data (Report No. MR-0033-1602). Research Triangle Park, NC: RTI Press. \url{http://www.rti.org/publication/visualization-categorical-longitudinal-and-times-series-data}
}
\seealso{
\code{\link{makePatterns}} for usage of the output from \code{norpt} as input.
}
\author{
Stephen Tueller
}
