% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longCat.R
\name{summary,longCat-method}
\alias{summary,longCat-method}
\title{Summary Method for longCat Objects}
\usage{
\S4method{summary}{longCat}(object, ...)
}
\arguments{
\item{object}{An object of class \code{longCat} to summarize.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
Prints a summary of the \code{longCat} object, including dimensions, the number of unique factors,
whether the data is sorted, group information, and possibly sorted group information. The function
itself invisibly returns \code{NULL}.
}
\description{
Provides a summary of an \code{longCat} object, including information about its dimensions,
factor details, sorting status, and group composition. If the object is sorted, additional
information relevant to sorting is also displayed.
}
