% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean.events}
\alias{clean.events}
\title{Clean Events Data}
\usage{
clean.events(events, event.times)
}
\arguments{
\item{events}{A data frame containing event data.}

\item{event.times}{A data frame containing the timestamps associated with each event in the `events` data. The dimensions (number of rows and columns) of `event.times` should match those of `events`.}
}
\value{
A list containing the cleaned `events` data frame and the cleaned `event.times` data frame.
}
\description{
Removes columns with only NA values from event data and corresponding event times.

This function iterates through each row of the `events` data and identifies columns containing only NA values. It then removes those columns from both the `events` and `event.times` data, while maintaining the corresponding row order.
}
