% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorter.R
\docType{data}
\name{example3}
\alias{example3}
\title{Simulated Data for Illustrating \code{\link{longCat}} and \code{\link{longCatPlot}}}
\format{
A data frame with 100 subjects and 6 time points. Each row represents a subject, and each column represents a time point.
}
\source{
Tueller, S. J., Van Dorn, R. A., & Bobashev, G. V. (2016). Visualization of categorical longitudinal and times series data (Report No. MR-0033-1602). Research Triangle Park, NC: RTI Press.
}
\usage{
data(example3)
}
\description{
Simulated data for illustrating \code{\link{longCat}} and \code{\link{longCatPlot}}.
}
\keyword{datasets}
