\name{example3}
\docType{data}
\alias{example3}
\title{Simulated Data for Illustrating longCat and longCatPlot}
\description{
Simulated data for illustrating \code{\link{longCat}} and \code{\link{longCatPlot}}.
}
\usage{example3}
\format{Data frame with 100 subjects and 6 time points}
\references{
Tueller, S. J., Van Dorn, R. A., & Bobashev, G. V. (2016). Visualization of categorical longitudinal and times series data (Report No. MR-0033-1602). Research Triangle Park, NC: RTI Press.
}
\keyword{datasets}
