\name{levelCheck}
\alias{levelCheck}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function to Check Factor Levels
}
\description{
Internal function of checking the levels of \code{y} called by \code{\link{longCat}}.
}
\usage{
levelCheck(y)
}
\arguments{
  \item{y}{a data matrix, see documentation for \code{y} in \code{\link{longCat}}.}
}
\references{
Tueller, S. J., Van Dorn, R. A., & Bobashev, G. V. (2016). Visualization of categorical longitudinal and times series data (Report No. MR-0033-1602). Research Triangle Park, NC: RTI Press. http://www.rti.org/publication/visualization-categorical-longitudinal-and-times-series-data}
\author{Stephen Tueller}

\seealso{
\code{\link{longCat}}.
}

