% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelCheck.R
\name{levelCheck}
\alias{levelCheck}
\title{Check Levels in a Data Frame or Matrix}
\usage{
levelCheck(y)
}
\arguments{
\item{y}{A data frame or matrix to check for levels of variables.}
}
\value{
Numeric vector of all unique values across all variables in `y`.
}
\description{
This function checks each variable in a data frame or matrix `y` for the number of unique levels.
If any variable contains 10 or more unique levels, a warning is issued suggesting consideration
of continuous methods. It returns the unique values across all variables as a numeric vector.
}
\examples{
y <- data.frame(a = 1:5, b = factor(c("yes", "no", "yes", "maybe", "no")))
levelCheck(y)

}
