% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longCatPlot.R
\name{longCatPlot}
\alias{longCatPlot}
\title{Plotting of \code{longCat} Objects}
\usage{
longCatPlot(
  lc,
  xlab = "Days",
  ylab = NULL,
  cols = NULL,
  colScheme = "heat",
  reverse = FALSE,
  lwd = 0.5,
  lcex = 1,
  llwd = 3,
  legendBuffer = 0.12,
  groupBuffer = 0,
  groupRotation = 90,
  gcex = 1,
  seg.len = 1,
  xlas = 0,
  xcex = 1,
  ecex = 0.5,
  event.col = 1,
  plot.events = TRUE,
  which.events = NULL,
  n.events = NULL,
  event.pch = NULL,
  texclude = NULL,
  sort = TRUE,
  which.state = NULL,
  nth.state = NULL,
  not.state = FALSE,
  which.event = NULL,
  nth.event = NULL,
  not.event = FALSE,
  ...
)
}
\arguments{
\item{lc}{An object of class \code{longCat} created by \code{\link{longCat}}.}

\item{xlab}{A label for the x-axis. Default is "Days".}

\item{ylab}{A label for the y-axis. Default is \code{NULL} which is changed to "Each Line Represents a Participant" with the sample size appended (\code{lc$dim[1]}).}

\item{cols}{A numeric or character list of colors. To use internal color schemes, use \code{colScheme}.}

\item{colScheme}{Select a color scheme. See \code{\link{colChoose}} for available options.}

\item{reverse}{Color schemes are applied from the lowest to highest level of categorical data in \code{lc$y} or \code{lc$y.sorted}. Set \code{reverse=TRUE} to reverse this. Default is \code{FALSE}.}

\item{lwd}{Set the width of horizontal lines. Default is .5.}

\item{lcex}{Character expansion factor for the legend text. Default is 1.}

\item{llwd}{Set the width of lines in the legend. Default is 3.}

\item{legendBuffer}{Set proportion of the plot to retain for legends, must be in [0,1]. Default is .12.}

\item{groupBuffer}{Similar to legendBuffer, but for group labels on the left side of the plot. Default is 0.}

\item{groupRotation}{Rotation of the labels to reduce the needed size of \code{groupBuffer}. Default is 90 degrees.}

\item{gcex}{Character expansion factor for group labels. Default is 1.}

\item{seg.len}{Length of lines in the upper legend. Default is 1.}

\item{xlas}{Applied to the x-axis when \code{tLabels} are provided. See \code{las} in \code{\link{par}}.}

\item{xcex}{Applied to the x-axis when \code{tLabels} are provided. Default is 1.}

\item{ecex}{Used to size the points used to plot event points if \code{events} is not \code{NULL}. Default is .5.}

\item{event.col}{Color for plotting event indicators.}

\item{plot.events}{Logical - should events be plotted? Default is \code{TRUE}.}

\item{which.events}{Numeric vector - which events should be plotted.}

\item{n.events}{How many events should be plotted, e.g., \code{n.events=3} plots the first three events for each participant.}

\item{event.pch}{What plotting characters should be used. See \code{\link{points}}.}

\item{texclude}{A vector indicating the range of \code{times} and \code{event.times} to be plotted.}

\item{sort}{Logical - should \code{longCatPlot} sort the data on the fly using intelligent defaults? Default is \code{TRUE}.}

\item{which.state}{See \code{\link{alignTime}}.}

\item{nth.state}{See \code{\link{alignTime}}.}

\item{not.state}{See \code{\link{alignTime}}.}

\item{which.event}{See \code{\link{alignTime}}.}

\item{nth.event}{See \code{\link{alignTime}}.}

\item{not.event}{See \code{\link{alignTime}}.}

\item{...}{Arguments to be passed to \code{\link{plot}}.}
}
\description{
Function to plot \code{longCat} objects created by \code{\link{longCat}}.
}
\examples{
# Illustrate longCatPlot with the legend outside the plot
par(mfrow=c(1,1), bg='cornsilk3', mar=c(5.1, 4.1, 4.1, 8.1), xpd=TRUE)
cols <- longCatPlot(
  longCat(example3),
  legendBuffer=0,
  main='Horizontal Line Plot')
legend(7.1, 100, legend=1:5, col=cols, lty=1, lwd=2)
par(bg='transparent', mar = c(5, 4, 4, 2) + 0.1, xpd=FALSE)

### visualizing multivariate data: 3 items at 4 time points
library(MASS)
Sigma <- matrix(.25, 12, 12)
diag(Sigma) <- 1
set.seed(9845)
mu <- rep(c(-.5, 0, .5), 4) + rnorm(12, 0, .25)
set.seed(539)
ymv <- apply(mvrnorm(n=100, mu=mu, Sigma = Sigma), 2, cut, breaks=c(-Inf, 0, Inf), labels=c(0,1))
apply(ymv, 2, table)
(items <- rep(1:3, 4))
(times <- sort(rep(1:4, 3)))
tLabels <- paste('Time', 1:4)
Labels <- paste('Item', 1:3)

# plot time points within items
par(mfrow=c(2,2), bg='cornsilk3', mar=c(5.1, 4.1, 4.1, 8.1), xpd=TRUE)

item1  <- longCat(y=ymv[,items==1], tLabels=tLabels)
cols <- longCatPlot(item1, ylab='', main='Item 1', legendBuffer=0, xlab="", xlas=2)
legend(length(unique(times))+.1, nrow(ymv), legend=0:1, col=cols, lty=1, lwd=2, title='Response')

item2 <- longCat(y=ymv[,items==2], tLabels=tLabels)
longCatPlot(item2, ylab='', main='Item 2', legendBuffer=0, xlab="", xlas=2)
legend(length(unique(times))+.1, nrow(ymv), legend=0:1, col=cols, lty=1, lwd=2, title='Response')

item3 <- longCat(y=ymv[,items==3], tLabels=tLabels)
longCatPlot(item3, ylab='', main='Item 3', legendBuffer=0, xlab="", xlas=2)
legend(length(unique(times))+.1, nrow(ymv), legend=0:1, col=cols, lty=1, lwd=2, title='Response')

par(bg='transparent', mar = c(5, 4, 4, 2) + 0.1, xpd=FALSE)

# plot items within time points
par(mfrow=c(2,2), bg='cornsilk3', mar=c(5.1, 4.1, 4.1, 8.1), xpd=TRUE)

time1  <- longCat(y=ymv[,times==1], tLabels=Labels)
cols <- longCatPlot(time1, ylab='', main='Time 1', legendBuffer=0, xlab="")
legend(length(unique(times))+.1, nrow(ymv), legend=0:1, col=cols, lty=1, lwd=2, title='Response')

time2  <- longCat(y=ymv[,times==2], tLabels=Labels)
cols <- longCatPlot(time2, ylab='', main='Time 2', legendBuffer=0, xlab="")
legend(length(unique(times))+.1, nrow(ymv), legend=0:1, col=cols, lty=1, lwd=2, title='Response')

time3  <- longCat(y=ymv[,times==3], tLabels=Labels)
cols <- longCatPlot(time3, ylab='', main='Time 3', legendBuffer=0, xlab="")
legend(length(unique(times))+.1, nrow(ymv), legend=0:1, col=cols, lty=1, lwd=2, title='Response')

time4  <- longCat(y=ymv[,times==4], tLabels=Labels)
cols <- longCatPlot(time4, ylab='', main='Time 4', legendBuffer=0, xlab="")
legend(length(unique(times))+.1, nrow(ymv), legend=0:1, col=cols, lty=1, lwd=2, title='Response')

par(mfrow=c(1,1), bg='transparent', mar = c(5, 4, 4, 2) + 0.1, xpd=FALSE)

\dontrun{
  # for data sets with many rows, writing directly to a file
  # is much faster and unaffected by device resizing, see ?pdf
  pdf('C:/mydir/mysubdir/myfile.pdf')
  par(bg='cornsilk3')
  longCatPlot(f3lc, main='Sorted', colScheme='heat', lwd=2)
  par(mfrow=c(1,1), bg='transparent')
  dev.off()
  # see ?jpeg for picture file options
}
}
\seealso{
\code{\link{longCat}}, for creating the object to be plotted.
}
