% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelCheck.R
\name{lunique}
\alias{lunique}
\title{Count Unique Non-NA Values}
\usage{
lunique(y)
}
\arguments{
\item{y}{Input vector, matrix, or data frame.}
}
\value{
Integer representing the number of unique non-NA values.
}
\description{
This function calculates the number of unique non-NA values in a vector, matrix, or data frame.
}
\examples{
vec <- c(1, 2, 3, NA, 3, 4, NA)
lunique(vec) # Returns 4
}
